requirements.txt
Flask==0.10.1
flask-cors==3.0.7
requests==2.19.1
Flask-WTF
Jinja2==2.5.5
Flask-DebugToolbar
pyyaml==3.13
Werkzeug==0.14.1
